Fix occasional disappearing content in CV items on UWP; fixes #4658
